Implemented Wrapper script for the website crawlers
Implemented Wrapper script for all the currently implemented website crawlers. The script gets the Adapter list as stored in the DB and cross-references them with the implemented ones in the WebCrawlers directory.

The wrapper script accepts up to 3 arguments:
   *  -a or --adapter: the name of the adapter, required argument
   *  -s or --searchTerm: the search query term, required argument
   *  -n or --numberResults: the number of results, optional argument that gets the value 10 by default if empty

The wrapper script executes the respective web crawler script according to the input adapter by passing to the script the arguments for the search term and the number of results.