Update recommendation module
The script handles the product recommendation as a two-state process. The initial state returns results according to scemantic similarity of the search term and the processed metadata of the Product table. The recalculation state, assumes that the user has provided feedback by rating and/or discarding the initial results as irrelevant. The provided feedback is used to predict the rating and relevance of the returned items that were not evaluated by the user, thus refining the recommendation.